services/trading_core/requirements.txt
redis==5.0.8
PyYAML==6.0.2
prometheus-client==0.20.0
uvloop==0.19.0; platform_system != "Windows"
httpx==0.27.2
